Refactor component includes and add new logo component
This commit refactors the existing component includes for better organization and removes unnecessary spaces between the includes. Also added is the inclusion of a new 'blocks/logo' component to provide support for the usage of a brand logo on the web pages. This will aid in better branding and recognition.