Add title styling to Sass stylesheet
A new 'title' block is added to our Sass stylesheet. This update was made to accommodate the changes in our website structure. 'title' styling includes attributes for color, font-family, and for 'H1' child element such as font-size, font-weight, line-height, and letter-spacing. The font-family was changed to '$ff-inter' to improve font performance and the readability of our website.