Update README file content
This commit replaced the current project build guide in the README.md file with a short description of a test task. The purpose of this change is to make the information more relevant in the context of a test task project rather than a detailed guide for a full project.